+++ spline_library/hermite/quintic/looping_quintic_hermite_spline.cpp
@@ -9,7 +9,7 @@ LoopingQuinticHermiteSpline::LoopingQuinticHermiteSpline(
         const std::vector<Vector3D> &points,
         const std::vector<Vector3D> &tangents,
         const std::vector<Vector3D> &curvatures,
-        float alpha
+        double alpha
         )
     :Spline(points)
 {
